new_installer: allow offline operations
git-svn-id: file:///home/toad/git-migration/temprepository/trunk/apps/new_installer@13185 67a373e5-eb02-0410-a15c-ee090a768436

+++ res/unix/bin/install_freenet-stable-latest.sh
@@ -4,7 +4,11 @@ INSTALL_PATH="${INSTALL_PATH:-$PWD}"
 
 cd "$INSTALL_PATH"
 
-echo "Downloading freenet-stable-latest.jar"
-java -jar bin/sha1test.jar freenet-stable-latest.jar "$INSTALL_PATH" &>/dev/null || exit 1 
+if test ! -e offline
+then
+	echo "Downloading freenet-stable-latest.jar"
+	java -jar bin/sha1test.jar freenet-stable-latest.jar "$INSTALL_PATH" &>/dev/null || exit 1 
+fi
+
 rm -f freenet.jar
 ln -sf freenet-stable-latest.jar freenet.jar 2>&1 >/dev/null
